restructure home page
