+++ front-end/src/components/home/EyeComponent.tsx
@@ -1,5 +1,4 @@
 import React, { useState, useEffect} from 'react';
-import './../HomePage.css';
 
 interface EyeComponentProps {
   headClicked: boolean;
